+++ plutus-numeric/test/size/Main.hs
@@ -3,7 +3,6 @@ module Main (main) where
 import Integer qualified
 import NatRatio qualified
 import Natural qualified
-import Rational qualified
 import Test.Tasty (defaultMain, testGroup)
 
 main :: IO ()
@@ -11,6 +10,5 @@ main =
   defaultMain . testGroup "Size checks" $
     [ testGroup "Natural" Natural.tests
     , testGroup "Integer" Integer.tests
-    , testGroup "Rational" Rational.tests
     , testGroup "NatRatio" NatRatio.tests
     ]