+++ tasty-plutus/test/Properties/Validator.hs
@@ -9,20 +9,19 @@ module Validator (tests) where
 import Prelude hiding (($), (&&), (*), (+), (==))
 
 import Ledger.Crypto (PubKeyHash)
-import Ledger.Typed.Scripts (
-  TypedValidator,
-  ValidatorTypes (type DatumType, type RedeemerType),
-  mkTypedValidator,
- )
 import Plutus.V1.Ledger.Value (Value)
 import Test.QuickCheck.Plutus.Instances ()
-import Test.Tasty (TestTree)
+import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.Plutus.Context (
   ContextBuilder,
   Purpose (ForSpending),
   paysToPubKey,
  )
-import Test.Tasty.Plutus.Script.Property (scriptProperty, scriptPropertyPass)
+import Test.Tasty.Plutus.Script.Property (
+  paramScriptProperty,
+  scriptProperty,
+  scriptPropertyPass,
+ )
 import Test.Tasty.Plutus.TestData (
   Generator (GenForSpending),
   Methodology (Methodology),
@@ -37,7 +36,12 @@ import Test.Tasty.Plutus.TestData (
   ),
   passIf,
  )
-import Test.Tasty.Plutus.WithScript (toTestValidator, withValidator)
+import Test.Tasty.Plutus.TestScript (
+  TestScript,
+  mkTestValidator,
+  toTestValidator,
+ )
+import Test.Tasty.Plutus.WithScript (withTestScript)
 import Test.Tasty.QuickCheck (
   Gen,
   arbitrary,
@@ -47,26 +51,35 @@ import Test.Tasty.QuickCheck (
 
 --------------------------------------------------------------------------------
 
+import Ledger.Address (PaymentPubKeyHash (unPaymentPubKeyHash))
 import Plutus.V1.Ledger.Contexts (ScriptContext)
+import PlutusTx (applyCode, liftCode)
 import PlutusTx.Prelude (traceIfFalse, ($), (&&), (*), (+), (==))
 import PlutusTx.TH (compile)
 import Wallet.Emulator.Types (WalletNumber (WalletNumber))
-import Wallet.Emulator.Wallet (fromWalletNumber, walletPubKeyHash)
+import Wallet.Emulator.Wallet (fromWalletNumber, mockWalletPaymentPubKeyHash)
 
 --------------------------------------------------------------------------------
 
 tests :: TestTree
 tests =
-  withValidator "simpleValidator:" typedSimpleValidator $ do
-    scriptProperty "Validator checks the sum of the inputs" $
-      GenForSpending gen1 transform1
-    scriptProperty "Validator checks the product of the inputs" $
-      GenForSpending gen1 transform2
-    scriptPropertyPass "Validator succeeds if the sum and product are correct" $
-      GenForSpending gen1 transform3
+  testGroup
+    "Validator tests"
+    [ withTestScript "With simple TestValidator:" simpleTestValidator $ do
+        scriptProperty "Validator checks the sum of the inputs" $
+          GenForSpending genForSimple transformForSimple1
+        scriptProperty "Validator checks the product of the inputs" $
+          GenForSpending genForSimple transformForSimple2
+        scriptPropertyPass "Validator succeeds if the sum and product are correct" $
+          GenForSpending genForSimple transformForSimple3
+    , paramScriptProperty
+        "Validator checks secret key"
+        (\(secret, _, _) -> paramTestValidator secret)
+        $ GenForSpending genForParam transformForParam
+    ]
 
-gen1 :: Methodology (Integer, Integer, Integer, Integer, Value)
-gen1 = Methodology gen' genericShrink
+genForSimple :: Methodology (Integer, Integer, Integer, Integer, Value)
+genForSimple = Methodology gen' genericShrink
   where
     gen' :: Gen (Integer, Integer, Integer, Integer, Value)
     gen' = do
@@ -75,8 +88,10 @@ gen1 = Methodology gen' genericShrink
       pure (i1, i2, iSum, iProd, val)
 
 -- | Creates TestItems with an arbitrary sum used in Redeemer
-transform1 :: (Integer, Integer, Integer, Integer, Value) -> TestItems ( 'ForSpending (Integer, Integer) (Integer, Integer))
-transform1 (i1, i2, iSum, _, val) =
+transformForSimple1 ::
+  (Integer, Integer, Integer, Integer, Value) ->
+  TestItems ( 'ForSpending (Integer, Integer) (Integer, Integer))
+transformForSimple1 (i1, i2, iSum, _, val) =
   ItemsForSpending
     { spendDatum = (i1, i2)
     , spendRedeemer = (iSum, i1 * i2)
@@ -91,8 +106,10 @@ transform1 (i1, i2, iSum, _, val) =
     out = if iSum == i1 + i2 then Pass else Fail
 
 -- | Creates TestItems with an arbitrary product used in Redeemer
-transform2 :: (Integer, Integer, Integer, Integer, Value) -> TestItems ( 'ForSpending (Integer, Integer) (Integer, Integer))
-transform2 (i1, i2, _, iProd, val) =
+transformForSimple2 ::
+  (Integer, Integer, Integer, Integer, Value) ->
+  TestItems ( 'ForSpending (Integer, Integer) (Integer, Integer))
+transformForSimple2 (i1, i2, _, iProd, val) =
   ItemsForSpending
     { spendDatum = (i1, i2)
     , spendRedeemer = (i1 + i2, iProd)
@@ -107,8 +124,10 @@ transform2 (i1, i2, _, iProd, val) =
     out = passIf $ iProd == i1 * i2
 
 -- | Always creates TestItems with correct sum and product
-transform3 :: (Integer, Integer, Integer, Integer, Value) -> TestItems ( 'ForSpending (Integer, Integer) (Integer, Integer))
-transform3 (i1, i2, _, _, val) =
+transformForSimple3 ::
+  (Integer, Integer, Integer, Integer, Value) ->
+  TestItems ( 'ForSpending (Integer, Integer) (Integer, Integer))
+transformForSimple3 (i1, i2, _, _, val) =
   ItemsForSpending
     { spendDatum = (i1, i2)
     , spendRedeemer = (i1 + i2, i1 * i2)
@@ -129,7 +148,8 @@ transform3 (i1, i2, _, _, val) =
   To spend some value, locked by the script with two integers
   you must provide the correct pair of sum and product of these integers.
 -}
-simpleValidator :: (Integer, Integer) -> (Integer, Integer) -> ScriptContext -> Bool
+simpleValidator ::
+  (Integer, Integer) -> (Integer, Integer) -> ScriptContext -> Bool
 simpleValidator (i1, i2) (iSum, iProd) _ = correctSum && correctProduct
   where
     correctSum :: Bool
@@ -142,17 +162,57 @@ simpleValidator (i1, i2) (iSum, iProd) _ = correctSum && correctProduct
       traceIfFalse "The product is wrong" $
         iProd == i1 * i2
 
-data TestScript
-
-instance ValidatorTypes TestScript where
-  type RedeemerType TestScript = (Integer, Integer)
-  type DatumType TestScript = (Integer, Integer)
-
-typedSimpleValidator :: TypedValidator TestScript
-typedSimpleValidator =
-  mkTypedValidator @TestScript
+simpleTestValidator ::
+  TestScript ( 'ForSpending (Integer, Integer) (Integer, Integer))
+simpleTestValidator =
+  mkTestValidator
     $$(compile [||simpleValidator||])
     $$(compile [||toTestValidator||])
 
+genForParam :: Methodology (Integer, Integer, Value)
+genForParam = Methodology gen genericShrink
+  where
+    gen :: Gen (Integer, Integer, Value)
+    gen = do
+      (i, val) <- arbitrary
+      i' <- oneof [pure i, arbitrary]
+      pure (i, i', val)
+
+-- | Creates TestItems with an arbitrary sum used in Redeemer
+transformForParam ::
+  (Integer, Integer, Value) ->
+  TestItems ( 'ForSpending () Integer)
+transformForParam (secret, guess, val) =
+  ItemsForSpending
+    { spendDatum = ()
+    , spendRedeemer = guess
+    , spendValue = val
+    , spendCB = cb
+    , spendOutcome = out
+    }
+  where
+    cb :: ContextBuilder ( 'ForSpending () Integer)
+    cb = paysToPubKey userPKHash val
+    out :: Outcome
+    out = if secret == guess then Pass else Fail
+
+paramValidator :: Integer -> () -> Integer -> ScriptContext -> Bool
+paramValidator secret _ guess _ = correctGuess
+  where
+    correctGuess :: Bool
+    correctGuess =
+      traceIfFalse "The guess is wrong" $
+        guess == secret
+
+paramTestValidator :: Integer -> TestScript ( 'ForSpending () Integer)
+paramTestValidator secret =
+  mkTestValidator
+    ($$(compile [||paramValidator||]) `applyCode` liftCode secret)
+    $$(compile [||toTestValidator||])
+
 userPKHash :: PubKeyHash
-userPKHash = walletPubKeyHash $ fromWalletNumber $ WalletNumber 1
+userPKHash =
+  unPaymentPubKeyHash
+    . mockWalletPaymentPubKeyHash
+    . fromWalletNumber
+    $ WalletNumber 1